Extract model crate
The extracted `model` crate still has some suspicious dependencies
such as `sqlx`, which we should strive to remove (in the future).
