+++ daemon-tests/Cargo.toml
@@ -12,6 +12,7 @@ bdk-ext = { path = "../bdk-ext" }
 daemon = { path = "../daemon" }
 mockall = "0.11"
 mockall_derive = "0.11"
+model = { path = "../model" }
 rand = "0.6"
 rust_decimal = "1.20"
 rust_decimal_macros = "1.21"
