+++ daemon-tests/src/lib.rs
@@ -10,15 +10,6 @@ use daemon::connection::connect;
 use daemon::connection::ConnectionStatus;
 use daemon::db;
 use daemon::maker_cfd;
-use daemon::model;
-use daemon::model::cfd::OrderId;
-use daemon::model::cfd::Role;
-use daemon::model::FundingRate;
-use daemon::model::Identity;
-use daemon::model::OpeningFee;
-use daemon::model::Price;
-use daemon::model::TxFeeRate;
-use daemon::model::Usd;
 use daemon::projection;
 use daemon::projection::Cfd;
 use daemon::projection::CfdOrder;
@@ -28,7 +19,15 @@ use daemon::seed::Seed;
 use daemon::MakerActorSystem;
 use daemon::HEARTBEAT_INTERVAL;
 use daemon::N_PAYOUTS;
-use daemon::SETTLEMENT_INTERVAL;
+use model::cfd::OrderId;
+use model::cfd::Role;
+use model::FundingRate;
+use model::Identity;
+use model::OpeningFee;
+use model::Price;
+use model::TxFeeRate;
+use model::Usd;
+use model::SETTLEMENT_INTERVAL;
 use rust_decimal::Decimal;
 use rust_decimal_macros::dec;
 use std::net::IpAddr;
