+++ daemon-tests/src/maia.rs
@@ -1,8 +1,8 @@
 use daemon::maia::secp256k1_zkp::schnorrsig;
 use daemon::maia::secp256k1_zkp::SecretKey;
-use daemon::olivia;
-use daemon::olivia::BitMexPriceEventId;
 use daemon::oracle;
+use model::olivia;
+use model::olivia::BitMexPriceEventId;
 use std::str::FromStr;
 
 #[allow(dead_code)]
@@ -52,8 +52,8 @@ impl OliviaData {
         }
     }
 
-    pub fn announcement(&self) -> oracle::Announcement {
-        oracle::Announcement {
+    pub fn announcement(&self) -> olivia::Announcement {
+        olivia::Announcement {
             id: self.id,
             expected_outcome_time: self.id.timestamp(),
             nonce_pks: self.nonce_pks.clone(),
