+++ daemon-tests/src/mocks/oracle.rs
@@ -1,9 +1,9 @@
 use crate::maia::OliviaData;
 use async_trait::async_trait;
-use daemon::olivia;
-use daemon::olivia::BitMexPriceEventId;
 use daemon::oracle;
 use mockall::*;
+use model::olivia;
+use model::olivia::BitMexPriceEventId;
 use std::sync::Arc;
 use time::OffsetDateTime;
 use tokio::sync::Mutex;
@@ -37,7 +37,7 @@ impl OracleActor {
     async fn handle(
         &mut self,
         msg: oracle::GetAnnouncement,
-    ) -> Result<oracle::Announcement, oracle::NoAnnouncement> {
+    ) -> Result<olivia::Announcement, oracle::NoAnnouncement> {
         self.mock.lock().await.get_announcement(msg)
     }
 
@@ -55,7 +55,7 @@ pub trait Oracle {
     fn get_announcement(
         &mut self,
         _msg: oracle::GetAnnouncement,
-    ) -> Result<oracle::Announcement, oracle::NoAnnouncement> {
+    ) -> Result<olivia::Announcement, oracle::NoAnnouncement> {
         unreachable!("mockall will reimplement this method")
     }
 
@@ -77,6 +77,7 @@ pub fn dummy_wrong_attestation() -> oracle::Attestation {
         price,
         scalars,
     } = OliviaData::example_0().attestation().into_inner();
+
     oracle::Attestation::new(olivia::Attestation {
         id: BitMexPriceEventId::with_20_digits(OffsetDateTime::now_utc()),
         price,
