+++ daemon-tests/src/mocks/wallet.rs
@@ -12,10 +12,10 @@ use daemon::bdk::FeeRate;
 use daemon::maia::secp256k1_zkp::Secp256k1;
 use daemon::maia::PartyParams;
 use daemon::maia::TxBuilderExt;
-use daemon::model::Timestamp;
-use daemon::model::WalletInfo;
 use daemon::wallet;
 use mockall::*;
+use model::Timestamp;
+use model::WalletInfo;
 use rand::thread_rng;
 use std::sync::Arc;
 use tokio::sync::Mutex;
