+++ daemon/src/collab_settlement_taker.rs
@@ -1,18 +1,18 @@
 use crate::cfd_actors::load_cfd;
 use crate::connection;
-use crate::model::cfd;
-use crate::model::cfd::CfdEvent;
-use crate::model::cfd::CollaborativeSettlement;
-use crate::model::cfd::CollaborativeSettlementCompleted;
-use crate::model::cfd::Completed;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::SettlementProposal;
-use crate::model::Price;
 use crate::process_manager;
 use crate::wire;
 use anyhow::anyhow;
 use anyhow::Result;
 use async_trait::async_trait;
+use model::cfd;
+use model::cfd::CfdEvent;
+use model::cfd::CollaborativeSettlement;
+use model::cfd::CollaborativeSettlementCompleted;
+use model::cfd::Completed;
+use model::cfd::OrderId;
+use model::cfd::SettlementProposal;
+use model::Price;
 use std::time::Duration;
 use tokio_tasks::Tasks;
 use xtra_productivity::xtra_productivity;
