+++ daemon/src/command.rs
@@ -1,10 +1,10 @@
 use crate::cfd_actors::load_cfd;
-use crate::model::cfd::Cfd;
-use crate::model::cfd::Event;
 use crate::process_manager;
 use crate::OrderId;
 use anyhow::Context;
 use anyhow::Result;
+use model::cfd::Cfd;
+use model::cfd::Event;
 use xtra::Address;
 
 #[derive(Clone)]
