+++ daemon/src/connection.rs
@@ -1,10 +1,5 @@
 use crate::collab_settlement_taker;
 use crate::future_ext::FutureExt;
-use crate::model::cfd::OrderId;
-use crate::model::Identity;
-use crate::model::Price;
-use crate::model::Timestamp;
-use crate::model::Usd;
 use crate::noise;
 use crate::rollover_taker;
 use crate::setup_taker;
@@ -21,6 +16,11 @@ use bdk::bitcoin::Amount;
 use futures::SinkExt;
 use futures::StreamExt;
 use futures::TryStreamExt;
+use model::cfd::OrderId;
+use model::Identity;
+use model::Price;
+use model::Timestamp;
+use model::Usd;
 use std::net::SocketAddr;
 use std::time::Duration;
 use std::time::SystemTime;
