+++ daemon/src/maker_cfd.rs
@@ -3,23 +3,6 @@ use crate::cfd_actors::insert_cfd_and_update_feed;
 use crate::collab_settlement_maker;
 use crate::command;
 use crate::maker_inc_connections;
-use crate::model::cfd::Cfd;
-use crate::model::cfd::CollaborativeSettlementCompleted;
-use crate::model::cfd::Order;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Origin;
-use crate::model::cfd::Role;
-use crate::model::cfd::RolloverCompleted;
-use crate::model::cfd::RolloverProposal;
-use crate::model::cfd::SettlementProposal;
-use crate::model::cfd::SetupCompleted;
-use crate::model::FundingRate;
-use crate::model::Identity;
-use crate::model::OpeningFee;
-use crate::model::Position;
-use crate::model::Price;
-use crate::model::TxFeeRate;
-use crate::model::Usd;
 use crate::oracle;
 use crate::process_manager;
 use crate::projection;
@@ -35,6 +18,23 @@ use anyhow::Context;
 use anyhow::Result;
 use async_trait::async_trait;
 use bdk::bitcoin::secp256k1::schnorrsig;
+use model::cfd::Cfd;
+use model::cfd::CollaborativeSettlementCompleted;
+use model::cfd::Order;
+use model::cfd::OrderId;
+use model::cfd::Origin;
+use model::cfd::Role;
+use model::cfd::RolloverCompleted;
+use model::cfd::RolloverProposal;
+use model::cfd::SettlementProposal;
+use model::cfd::SetupCompleted;
+use model::FundingRate;
+use model::Identity;
+use model::OpeningFee;
+use model::Position;
+use model::Price;
+use model::TxFeeRate;
+use model::Usd;
 use std::collections::HashSet;
 use time::Duration;
 use tokio_tasks::Tasks;
