+++ daemon/src/maker_inc_connections.rs
@@ -4,9 +4,6 @@ use crate::maker_cfd;
 use crate::maker_cfd::FromTaker;
 use crate::maker_cfd::TakerConnected;
 use crate::maker_cfd::TakerDisconnected;
-use crate::model::cfd::Order;
-use crate::model::cfd::OrderId;
-use crate::model::Identity;
 use crate::noise;
 use crate::noise::TransportStateExt;
 use crate::rollover_maker;
@@ -24,6 +21,9 @@ use async_trait::async_trait;
 use futures::SinkExt;
 use futures::StreamExt;
 use futures::TryStreamExt;
+use model::cfd::Order;
+use model::cfd::OrderId;
+use model::Identity;
 use std::collections::HashMap;
 use std::net::SocketAddr;
 use std::time::Duration;
