+++ daemon/src/oracle.rs
@@ -1,13 +1,13 @@
 use crate::db;
-use crate::model::cfd::CfdEvent;
-use crate::model::cfd::Event;
-use crate::olivia;
-use crate::olivia::BitMexPriceEventId;
 use crate::try_continue;
 use anyhow::Context;
 use anyhow::Result;
 use async_trait::async_trait;
 use maia::secp256k1_zkp::schnorrsig;
+use model::cfd::CfdEvent;
+use model::cfd::Event;
+use model::olivia;
+use model::olivia::BitMexPriceEventId;
 use sqlx::SqlitePool;
 use std::collections::HashMap;
 use std::collections::HashSet;
