+++ daemon/src/projection.rs
@@ -1,27 +1,4 @@
 use crate::db;
-use crate::model;
-use crate::model::calculate_funding_fee;
-use crate::model::cfd::calculate_long_liquidation_price;
-use crate::model::cfd::calculate_long_margin;
-use crate::model::cfd::calculate_profit;
-use crate::model::cfd::calculate_profit_at_price;
-use crate::model::cfd::calculate_short_margin;
-use crate::model::cfd::CfdEvent;
-use crate::model::cfd::Dlc;
-use crate::model::cfd::Event;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Origin;
-use crate::model::cfd::Role;
-use crate::model::FeeAccount;
-use crate::model::FundingRate;
-use crate::model::Identity;
-use crate::model::Leverage;
-use crate::model::Position;
-use crate::model::Price;
-use crate::model::Timestamp;
-use crate::model::TradingPair;
-use crate::model::Usd;
-use crate::model::SETTLEMENT_INTERVAL;
 use crate::Order;
 use anyhow::Context;
 use anyhow::Result;
@@ -35,6 +12,28 @@ use bdk::bitcoin::Txid;
 use bdk::miniscript::DescriptorTrait;
 use core::fmt;
 use maia::TransactionExt;
+use model::calculate_funding_fee;
+use model::cfd::calculate_long_liquidation_price;
+use model::cfd::calculate_long_margin;
+use model::cfd::calculate_profit;
+use model::cfd::calculate_profit_at_price;
+use model::cfd::calculate_short_margin;
+use model::cfd::CfdEvent;
+use model::cfd::Dlc;
+use model::cfd::Event;
+use model::cfd::OrderId;
+use model::cfd::Origin;
+use model::cfd::Role;
+use model::FeeAccount;
+use model::FundingRate;
+use model::Identity;
+use model::Leverage;
+use model::Position;
+use model::Price;
+use model::Timestamp;
+use model::TradingPair;
+use model::Usd;
+use model::SETTLEMENT_INTERVAL;
 use parse_display::Display;
 use parse_display::FromStr;
 use rust_decimal::Decimal;
