+++ daemon/src/rollover_maker.rs
@@ -2,15 +2,6 @@ use crate::command;
 use crate::maker_inc_connections;
 use crate::maker_inc_connections::RegisterRollover;
 use crate::maker_inc_connections::TakerMessage;
-use crate::model::cfd::Completed;
-use crate::model::cfd::Dlc;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Role;
-use crate::model::cfd::RolloverCompleted;
-use crate::model::FundingFee;
-use crate::model::FundingRate;
-use crate::model::Identity;
-use crate::model::TxFeeRate;
 use crate::oracle;
 use crate::oracle::GetAnnouncement;
 use crate::process_manager;
@@ -26,6 +17,15 @@ use futures::channel::mpsc;
 use futures::channel::mpsc::UnboundedSender;
 use futures::future;
 use futures::SinkExt;
+use model::cfd::Completed;
+use model::cfd::Dlc;
+use model::cfd::OrderId;
+use model::cfd::Role;
+use model::cfd::RolloverCompleted;
+use model::FundingFee;
+use model::FundingRate;
+use model::Identity;
+use model::TxFeeRate;
 use tokio_tasks::Tasks;
 use xtra::prelude::MessageChannel;
 use xtra::Context;
