+++ daemon/src/rollover_taker.rs
@@ -1,14 +1,5 @@
 use crate::command;
 use crate::connection;
-use crate::model::cfd::Dlc;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Role;
-use crate::model::cfd::RolloverCompleted;
-use crate::model::FundingFee;
-use crate::model::FundingRate;
-use crate::model::Timestamp;
-use crate::model::TxFeeRate;
-use crate::olivia::BitMexPriceEventId;
 use crate::oracle;
 use crate::oracle::GetAnnouncement;
 use crate::process_manager;
@@ -24,6 +15,15 @@ use futures::channel::mpsc::UnboundedSender;
 use futures::future;
 use futures::SinkExt;
 use maia::secp256k1_zkp::schnorrsig;
+use model::cfd::Dlc;
+use model::cfd::OrderId;
+use model::cfd::Role;
+use model::cfd::RolloverCompleted;
+use model::olivia::BitMexPriceEventId;
+use model::FundingFee;
+use model::FundingRate;
+use model::Timestamp;
+use model::TxFeeRate;
 use std::time::Duration;
 use tokio_tasks::Tasks;
 use xtra::prelude::MessageChannel;
