+++ daemon/src/setup_contract.rs
@@ -1,13 +1,4 @@
 use crate::future_ext::FutureExt;
-use crate::model::cfd::Cet;
-use crate::model::cfd::Dlc;
-use crate::model::cfd::RevokedCommit;
-use crate::model::cfd::Role;
-use crate::model::cfd::RolloverParams;
-use crate::model::cfd::SetupParams;
-use crate::model::cfd::CET_TIMELOCK;
-use crate::olivia;
-use crate::payout_curve;
 use crate::transaction_ext::TransactionExt;
 use crate::wallet;
 use crate::wire::Msg0;
@@ -48,6 +39,15 @@ use maia::spending_tx_sighash;
 use maia::Announcement;
 use maia::PartyParams;
 use maia::PunishParams;
+use model::cfd::Cet;
+use model::cfd::Dlc;
+use model::cfd::RevokedCommit;
+use model::cfd::Role;
+use model::cfd::RolloverParams;
+use model::cfd::SetupParams;
+use model::cfd::CET_TIMELOCK;
+use model::olivia;
+use model::payout_curve;
 use std::collections::HashMap;
 use std::iter::FromIterator;
 use std::ops::RangeInclusive;
