+++ daemon/src/setup_maker.rs
@@ -1,14 +1,6 @@
 use crate::command;
 use crate::maker_inc_connections;
 use crate::maker_inc_connections::TakerMessage;
-use crate::model::cfd::Dlc;
-use crate::model::cfd::Order;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Role;
-use crate::model::cfd::SetupCompleted;
-use crate::model::Identity;
-use crate::model::Usd;
-use crate::olivia::Announcement;
 use crate::process_manager;
 use crate::setup_contract;
 use crate::wallet;
@@ -23,6 +15,14 @@ use futures::channel::mpsc::UnboundedSender;
 use futures::future;
 use futures::SinkExt;
 use maia::secp256k1_zkp::schnorrsig;
+use model::cfd::Dlc;
+use model::cfd::Order;
+use model::cfd::OrderId;
+use model::cfd::Role;
+use model::cfd::SetupCompleted;
+use model::olivia::Announcement;
+use model::Identity;
+use model::Usd;
 use tokio_tasks::Tasks;
 use xtra::prelude::MessageChannel;
 use xtra::Address;
