+++ daemon/src/setup_taker.rs
@@ -1,11 +1,5 @@
 use crate::command;
 use crate::connection;
-use crate::model::cfd::Dlc;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Role;
-use crate::model::cfd::SetupCompleted;
-use crate::model::Usd;
-use crate::olivia::Announcement;
 use crate::process_manager;
 use crate::setup_contract;
 use crate::wallet;
@@ -20,6 +14,12 @@ use futures::channel::mpsc::UnboundedSender;
 use futures::future;
 use futures::SinkExt;
 use maia::secp256k1_zkp::schnorrsig;
+use model::cfd::Dlc;
+use model::cfd::OrderId;
+use model::cfd::Role;
+use model::cfd::SetupCompleted;
+use model::olivia::Announcement;
+use model::Usd;
 use std::time::Duration;
 use tokio_tasks::Tasks;
 use xtra::prelude::*;
