+++ daemon/src/taker_cfd.rs
@@ -2,15 +2,6 @@ use crate::cfd_actors;
 use crate::cfd_actors::insert_cfd_and_update_feed;
 use crate::collab_settlement_taker;
 use crate::connection;
-use crate::model::cfd::Cfd;
-use crate::model::cfd::Order;
-use crate::model::cfd::OrderId;
-use crate::model::cfd::Origin;
-use crate::model::cfd::Role;
-use crate::model::Identity;
-use crate::model::Position;
-use crate::model::Price;
-use crate::model::Usd;
 use crate::oracle;
 use crate::process_manager;
 use crate::projection;
@@ -20,6 +11,15 @@ use anyhow::Context as _;
 use anyhow::Result;
 use async_trait::async_trait;
 use bdk::bitcoin::secp256k1::schnorrsig;
+use model::cfd::Cfd;
+use model::cfd::Order;
+use model::cfd::OrderId;
+use model::cfd::Origin;
+use model::cfd::Role;
+use model::Identity;
+use model::Position;
+use model::Price;
+use model::Usd;
 use tokio_tasks::Tasks;
 use xtra::prelude::*;
 use xtra::Actor as _;
