+++ daemon/src/wire.rs
@@ -1,10 +1,3 @@
-use crate::model::cfd::Order;
-use crate::model::cfd::OrderId;
-use crate::model::FundingRate;
-use crate::model::Price;
-use crate::model::Timestamp;
-use crate::model::TxFeeRate;
-use crate::model::Usd;
 use crate::noise::NOISE_MAX_MSG_LEN;
 use crate::noise::NOISE_TAG_LEN;
 use crate::olivia::BitMexPriceEventId;
@@ -23,6 +16,13 @@ use maia::secp256k1_zkp::SecretKey;
 use maia::CfdTransactions;
 use maia::PartyParams;
 use maia::PunishParams;
+use model::cfd::Order;
+use model::cfd::OrderId;
+use model::FundingRate;
+use model::Price;
+use model::Timestamp;
+use model::TxFeeRate;
+use model::Usd;
 use serde::de::DeserializeOwned;
 use serde::Deserialize;
 use serde::Serialize;
