+++ maker/Cargo.toml
@@ -11,6 +11,7 @@ clap = { version = "3", features = ["derive"] }
 daemon = { path = "../daemon" }
 hex = "0.4"
 http-api-problem = { version = "0.51.0", features = ["rocket"] }
+model = { path = "../model" }
 rocket = { version = "0.5.0-rc.1", features = ["json", "uuid"] }
 rocket-basicauth = { path = "../rocket-basicauth" }
 rust-embed = "6.3"
