+++ maker/src/main.rs
@@ -7,9 +7,7 @@ use daemon::bdk::bitcoin;
 use daemon::bdk::bitcoin::Amount;
 use daemon::bdk::FeeRate;
 use daemon::db;
-use daemon::model::cfd::Role;
 use daemon::monitor;
-use daemon::olivia;
 use daemon::oracle;
 use daemon::projection;
 use daemon::seed::RandomSeed;
@@ -18,7 +16,9 @@ use daemon::wallet;
 use daemon::MakerActorSystem;
 use daemon::HEARTBEAT_INTERVAL;
 use daemon::N_PAYOUTS;
-use daemon::SETTLEMENT_INTERVAL;
+use model::cfd::Role;
+use model::olivia;
+use model::SETTLEMENT_INTERVAL;
 use rocket::fairing::AdHoc;
 use shared_bin::logger;
 use shared_bin::logger::LevelFilter;
