+++ model/src/payout_curve/README.md
