+++ model/src/payout_curve/basis.rs
