+++ model/src/payout_curve/basis_eval.rs
