+++ model/src/payout_curve/compat.rs
