+++ model/src/payout_curve/csr_tools.rs
