+++ model/src/payout_curve/curve.rs
