+++ model/src/payout_curve/curve_factory.rs
