+++ model/src/payout_curve/splineobject.rs
