+++ model/src/payout_curve/utils.rs
