+++ shared-bin/Cargo.toml
@@ -10,6 +10,7 @@ description = "Code that is shared between the daemons but application specific 
 anyhow = "1"
 atty = "0.2"
 daemon = { path = "../daemon" }
+model = { path = "../model" }
 rocket = { version = "0.5.0-rc.1", features = ["json"] }
 serde = { version = "1", features = ["derive"] }
 time = "0.3"
