+++ taker/Cargo.toml
@@ -12,6 +12,7 @@ daemon = { path = "../daemon" }
 hex = "0.4"
 http-api-problem = { version = "0.51.0", features = ["rocket"] }
 itertools = "0.10"
+model = { path = "../model" }
 rocket = { version = "0.5.0-rc.1", features = ["json", "uuid"] }
 rocket-basicauth = { path = "../rocket-basicauth" }
 rust-embed = "6.3"
