+++ taker/src/main.rs
@@ -8,10 +8,7 @@ use daemon::bdk::bitcoin::Amount;
 use daemon::bdk::FeeRate;
 use daemon::connection::connect;
 use daemon::db;
-use daemon::model::cfd::Role;
-use daemon::model::Identity;
 use daemon::monitor;
-use daemon::olivia;
 use daemon::oracle;
 use daemon::projection;
 use daemon::seed::RandomSeed;
@@ -21,7 +18,10 @@ use daemon::wallet;
 use daemon::TakerActorSystem;
 use daemon::HEARTBEAT_INTERVAL;
 use daemon::N_PAYOUTS;
-use daemon::SETTLEMENT_INTERVAL;
+use model::cfd::Role;
+use model::olivia;
+use model::Identity;
+use model::SETTLEMENT_INTERVAL;
 use rocket::fairing::AdHoc;
 use shared_bin::logger;
 use shared_bin::logger::LevelFilter;
