+++ taker/src/routes.rs
@@ -2,12 +2,6 @@ use daemon::bdk;
 use daemon::bdk::bitcoin::Amount;
 use daemon::bdk::bitcoin::Network;
 use daemon::connection::ConnectionStatus;
-use daemon::model::cfd::OrderId;
-use daemon::model::Leverage;
-use daemon::model::Price;
-use daemon::model::Timestamp;
-use daemon::model::Usd;
-use daemon::model::WalletInfo;
 use daemon::oracle;
 use daemon::projection;
 use daemon::projection::CfdAction;
@@ -16,6 +10,12 @@ use daemon::wallet;
 use daemon::TakerActorSystem;
 use http_api_problem::HttpApiProblem;
 use http_api_problem::StatusCode;
+use model::cfd::OrderId;
+use model::Leverage;
+use model::Price;
+use model::Timestamp;
+use model::Usd;
+use model::WalletInfo;
 use rocket::http::ContentType;
 use rocket::http::Status;
 use rocket::response::stream::Event;
